requirements.txt
loguru==0.7.0
pandas==2.0.2
psycopg2_binary==2.9.6
sshtunnel==0.4.0
